refactor: [19] add extensions
