+++ Sources/HealthChecks/PostgresHealthChecks/PostgresHealthChecks.swift
@@ -97,40 +97,3 @@ public struct PostgresHealthChecks: PostgresHealthChecksProtocol {
         return result
     }
 }
-
-protocol PsqlSendable {
-    func getVersionDescription() async throws -> String
-}
-
-extension Application {
-    /// `StorageKey` for KafkaSendable
-    private struct PsqlRequestKey: StorageKey {
-        /// KafkaSendable
-        typealias Value = PsqlSendable
-    }
-
-    /// Computed property for `KafkaSendable`
-    var psqlRequest: PsqlSendable? {
-        get { storage[PsqlRequestKey.self] }
-        set { storage[PsqlRequestKey.self] = newValue }
-    }
-}
-
-struct PsqlRequest: PsqlSendable {
-    let app: Application
-    func getVersionDescription() async throws -> String {
-        let rows = try? await (app.db(.psql) as? PostgresDatabase)?.simpleQuery("SELECT version()").get()
-        let row = rows?.first?.makeRandomAccess()
-        var connectionDescription = "ERROR: No connect to Postgres database"
-        if let result = (row?[data: "version"].string) {
-            connectionDescription = result
-        }
-        return connectionDescription
-    }
-}
-
-struct MockPsqlRequest: PsqlSendable {
-    func getVersionDescription() async throws -> String {
-        return "PostgreSQL 14.10 on x86_64-pc-linux-musl, compiled by gcc (Alpine 13.2.1_git20231014) 13.2.1 20231014, 64-bit"
-    }
-}
